fix: Improve atomicity of `LinkedItemList`s (#2456)
## What

Fix issues with consistency of the `SEGMENT_METAS` list in the presence of physical replication.

## Why

Physical replicas replay the WAL without regard for locks held on the primary. Consequently, there is no lock that we can hold that will prevent a physical replica from consuming a snapshot of the `SegmentMetaEntry`s immediately before one of the segment's files is deleted.

Additionally, two other bugs were causing the `SegmentMetaEntry`s to appear non-atomically on physical replicas: a bug in the `atomically` method, and the lack of hand-over-hand-locking in `LinkedList` implementations.

## How

* Split `LinkedItemList::retain` out of `LinkedItemList::garbage_collect` to provide different selection logic when we're deciding whether something can actually be freed.
* Add a `segment_metas_garbage` list, which contains segments which are no longer visible on a physical replication primary, but which might still be visible on a replica.
* Split out `impl Write for LinkedBytesListWriter` in order to simplify metadata management in `LinkedBytesList`.
* Execute hand-over-hand locking in `LinkedList` implementations to fix segment-corruption issues when partially written or freed pages are accessed.
* Fix issue in `atomically` where the previous content of a list might be freed before the header was flushed.

## Tests

`stressgres`.