feat: remove snapshot visibility for `SegmentMetaEntry`
In practice a segment is either visible or it's been deleted by a merge.
This changes the code so that's the case and no longer validates a
segment's visibility against the current snapshot.

There's a number of problems doing it the way we did which include:

- defers merging segments-created-in-this-transaction with
  segments-created-in-other-transactions, allowing the segment list to
  grow that much faster
- complicates logic around when a segment should be considered "visible"
  to a transaction or not, requiring us to track a Snapshot, which is
  dubious in some situations such as logical replication

This removes `SegmentMetaEntry`'s `xmin` field (turning it into
`_unused`) and essentially treating `xmax` as a boolean -- if it's 2
(`pg_sys::FrozenTransactionId`) then it's been deleted and we can
garbage collect it, and also ignore it during searching.  Otherwise,
it's visible and can't be gc'd.

Similarly, we adjust the `MergeEntry` to only contain an `xmin` (and
rename its `xmax` field to `_unused`) as for it we only care about the
status of the transaction doing the merge the entry represents.