chore: Serialize private planning data with serde (#2477)
## What

We currently serialize the `PrivateData` of planning using Postgres's `List` and `Node` types. But that data is never actually introspected by Postgres, so we can use whichever serialization method we want.

## Why

Switching to `serde` allows us to more easily introduce new data passed between planning and execution. It is never persisted, so backwards compatibility is not a concern.

#2453 in particular would like to adjust what is computed during planning vs execution, which will involve adding fields to the `PrivateData`.