chore: Move the selection of the custom scan execution method to planning (#2492)
# Ticket(s) Closed

- Closes #2453

## What

Moves the choice of execution method to the planning phase, so that we can directly consume the chosen method to make assertions about the properties of the output.

## Why

As described on #2453: the current split can cause the type of confusion exemplified by #2452, where although a property like `sort_direction` is set on the builder, whether it is actually _used_ is not known until an execution method is chosen. For example: the `TopN` execution mode will consume the `sort_direction` and `limit`, but other execution modes will not apply a sort.

## How

Introduce `enum ExecMethodType`, which is chosen during planning and then consumed verbatim during execution to construct the `ExecMethod`. 

## Tests

Existing tests.