fix: issue #2564 -- `NormalScanExecState` now properly resets on rescans (#2595)
`NormalScanExecState` was failing to reset enough of its internal state to realize that during a rescan it should actually execute the query again.

It's worth noting that the bug itself, despite the report in #2564, is not dependent on `ANALYZE` so much as post-ANALYZE, postgres would choose an entirely different query plan which didn't involve rescans, so this bug wasn't tickled.