perf: move `MergeLock` to its own buffer, allowing share access to metadata page (#2609)
# Ticket(s) Closed

- Closes #

## What

The `MERGE_LOCK` buffer was being used both as an exclusive lock and to
store metadata, which meant that in order to read metadata we needed an
exclusive lock on the buffer.

The merge lock has been moved to its own buffer, allowing concurrent
access to the metadata page. Along the way I've cleaned up the code a
bit too.

## Why

## How

## Tests
Existing tests/stressgres pass