perf: optimize `search_with_query_input` (#2608)
# Ticket(s) Closed

- Closes #

## What

Optimize the `search_with_query_input` implementation to remove as much
overhead as possible. When used, this function is called for every row
in expensive scans like sequential scans or as the filter on certain
join scans.

## Why

The less overhead the better.

## How

## Tests

This came out of investigating the performance of a query our qgen
proptests happened to generate.