feat: always do pushdowns when we can (#2639)
Until now, we would short-circuit out of our custom scan if the query
didn't contain at least one `@@@` usage somewhere in the query, even if
other quals, such as `field = 'value'` could be pushed down. Essentially
this was a measure to only do pushdowns when the user made it clear they
wanted/needed a custom scan.

This removes that short-circuit and will now send any query through our
custom scan where every qual is either directly `@@@` or is otherwise
pushdown-able.

## Why

We have enough test coverage now where we can confidently do this and
ensure that the custom scan plan will produce the same results as the
non-custom scan plan we'd have executed prior to this PR.

## How

## Tests

Two of the regression tests have changed plans b/c now they go through
the custom scan path. A third adds a new query to validate, which was
the result of a bug this change uncovered related to confusing which
relation a Var came from.

Additionally, we add some more logic on the qgen proptests to force
Postgres to do various plans to make sure we always get the correct
results.