perf: improve INSERT/UPDATE perf with `SerialIndexWriter` (#2658)
We now use the `SerialIndexWriter` to insert rows during `aminsert()`,
which seems to essentially double INSERT/UPDATE performance of small,
single-row statements.

This then led to `SearchIndexWriter` being renamed to
`SearchIndexDeleter` because now it's only used by `ambulkdelete()`.

We're basically one step away from being able to completely delete the
`ChannelDirectory` and all the complexity it contains!

As a result of all this, the `paradedb.statement_parallelism` GUC is now
gone. Internally we keep the concept, but it's hardcoded to 1 --
Postgres doesn't support parallel worker-based INSERT/UPDATE/COPY
statements.