chore: Clean up schema code (#2660)
# Ticket(s) Closed

- Closes #

## What

A refactor of our schema-related code. Most of the actual changes are in
`postgres/options.rs` and `schema/`. The rest of the diffs are just the
fallout of that.

As part of the cleanup I caught and fixed some things:

- The range field `fast` option can default to `false`, causing it to
not be sortable -- this should never happen
- We now use `serde` to reject any unknown fields in the `CREATE INDEX`
JSON config options
- Because unknown fields were previously not rejected, some of our unit
tests were asserting that the wrong configs were being deserialized --
this has been fixed
- We were allowing users to alias non-text fields -- this doesn't make
sense and the option has been removed

## Why

## How

- Deleted unnecessary abstractions like `SearchDocument` and
`SearchFieldId`
- Deleted confusing traits like `AsFieldType`
- Refactored `options.rs` and `schema/mod.rs`

## Tests
All existing tests should pass