fix: should be able to construct a `SearchField` for `ctid` (#2662)
## What

Like the title says, we also need to be able to construct a
`SearchField` instance for the postgres "ctid" system column. This was a
minor regression from #2660, because now we only rely on the index's
TupleDescriptor to lookup attributes.

## Why

Specifically, the `paradedb.find_ctid()` was broken due to this, which
is something used in some stressgres test suites.

## How

## Tests

Yes, yes there's a new one to assert that `paradedb.find_ctid()` works.