fix: Ensure that Top-N early exits when the limit has been reached (#2671)
## What

When an `ORDER BY` tiebreaker was used, `Top-N` (post #2624) would
iterate through all matches for a query (without emitting more than the
`LIMIT`) before exiting, because the `ExecMethod` would re-query when it
encountered `ExecState::Eof`, and `TopNScanExecState::query` was not
defending against that case.

Without an `ORDER BY` tiebreaker, a `Limit` node in the plan causes
Postgres to stop requesting tuples on its own: but with a tiebreaker,
our parent node (an `Incremental Sort` in this case) will continue
requesting tuples until our custom scan node indicates that there aren't
any more.

## How

Add an additional termination condition in `query`.

Additionally, replace (and fix) the `retry_count` metric with a more
general `query_count` metric.

## Tests

This was not originally caught in unit tests due to queries over small
dataset sizes eventually actually completing correctly.

Added a test that asserts that the expected number of queries are
executed.

---------

Co-authored-by: Philippe Noël <21990816+philippemnoel@users.noreply.github.com>
Co-authored-by: Eric Ridge <eebbrr@gmail.com>
Co-authored-by: Ming <ming.ying.nyc@gmail.com>
Co-authored-by: Moe <mdashti@gmail.com>