fix: stop TopN parallel workers from aggressively stealing all segments
Due to the nature of how Postgres launches and executes parallel scan
workers, it was possible for one (or a few) of the workers launched to
steal all (or most) of the segments, effectively defeating the
parallelism.

Address this by having the leader process write the actual number of
launched workers to the shared memory parallel state and have the
workers wait for that value to become available.  This only works if
`parallel_leader_participation` is on.

If it's off then the parallel workers take a different path and wait for
15ms (a number I made up) for as many workers to start as possible...
the hope being that that number is equal to the number of workers we
requested.

This also changes the segment checkout approach for `TopNExecState` such
that each worker picks the total number of segments it's supposed to
have up-front, rather than iteratively.  This leads to a bit of code
cleanup, removing the complexity around the `RefCell` that used to track
the their `claimed_segments`.