docs: Clarify message for enterprise (#2693)
# Ticket(s) Closed

- Closes #N/A

## What
Suggested by a user, and I thought it was a good idea. This clarifies
that you can't have more than 1 instance under ParadeDB Community, for
people not familiar with the implications of no WAL replication, and
adds a note for non-profits who might want access to ParadeDB
Enterprise.

## Why
^

## How
^

## Tests
N/A

---------

Signed-off-by: Philippe Noël <21990816+philippemnoel@users.noreply.github.com>