perf: sort segments by num docs for parallel workers (#2698)
## What

When segments are claimed by workers they're claimed from back-to-front
and our goal is to have the largest segments claimed first so that the
processing done on them takes longer, allowing more workers to checkout
their own segments.

This came about after some offline discussions with @stuhood.

## Why

We've seen situations where if an index's segments are very unbalanced
some parallel workers might not do any work because other workers
already did all the work. While that may be okay, it's probably better
to try and convince the system to be a bit more fair, and sorting
segments by num docs can allow for that.

## How

## Tests

Existing tests pass.