feat: `CREATE INDEX`/`REINDEX` exactly targets the CPU count by efficiently merging (#2681)
# Ticket(s) Closed

- Closes #

## What

- Allows `CREATE INDEX` to produce the optimal number of segments on its
own, without the user needing to run `force_merge`.
- With this change, the amount of indexing memory no longer limits how
many segments are created. As such, we can use the Postgres
`maintenance_work_mem` and `work_mem` settings for index build and
statements, respectively.

## Why

## How

- `SerialIndexWriter` now has the ability to produce an exact number of
`n` segments.
- `SerialIndexWriter` achieves this by first producing `n` segments. The
remaining segments are created in memory (backed by a `RamDirectory`)
and merged with the smallest segment. To guarantee an even distribution
of docs, the caller of `SearchIndexWriter` can specify
`min_docs_per_segment`.
- After launching parallel workers, the leader determines how many
segments each worker should target. This list of target segments is
written to the shared `WorkerCoordination` state. Then, each worker
"claims" a target segment count and passes it to the writer.
- After every segment is written, the writer garbage collect to
efficiently return space to the FSM.


## Tests
- Added unit tests to `SearchIndexWriter`
- Add regression tests

---------

Signed-off-by: Ming <ming.ying.nyc@gmail.com>
Co-authored-by: Eric Ridge <eebbrr@gmail.com>
Co-authored-by: Eric B. Ridge <eebbrr@paradedb.com>