test: Fix TopN for ascending string order, and add generated paging tests (#2635)
The tantivy `TopNComputer` was broken for ascending orders, but it was
only actually used for ascending orders (for the first time?) by
https://github.com/paradedb/tantivy/pull/45.

Expands the proptests to add a test of subqueries, and to add "paging"
expressions (`ORDER BY`, `LIMIT`, `OFFSET`) to the single query and
subquery tests. One of these tests reproduces the issue fixed by
https://github.com/paradedb/tantivy/pull/51.

To increase test coverage: particularly around top-n, and subqueries
(which seemingly hit re-scan cases more frequently). And to fix
https://github.com/paradedb/tantivy/pull/51.