fix: `paradedb.aggregate()` should account for `parallel_leader_participation` (#2699)
## What

If the leader is participating we should request 1 less parallel worker.

Also adds some debug statements about the number of workers requested
and timings for each worker.

## Why

This could have led to having 1 more worker than there are segments,
which would create an unbalanced amount of work in each worker, causing
performance to be much worse.

## How

## Tests

Existing tests pass. It's a little difficult to somehow assert, via a
test, that only a specific number of workers were started.