fix: Allow UUID columns to be used in fast field execution (#2707)
# Ticket(s) Closed

- Closes #2703

## What

As described in #2703, UUID columns which are indexed as `fast` are
incorrectly interpreted as numeric fast fields during
`collect_fast_fields`. This fails at execution time for `Mixed` and
`String` fast field execution modes.

Relatedly: a `UUID` `key_field` was not correctly being configured as
`fast` at index creation time, meaning that scans using it would not use
the fast field execution methods.

## Why

To fix the crash described on #2703, and to allow fast-field execution
to be used in more cases.

## How

In two commits:
1. fix the `FastFieldType` which is extracted for a UUID
2. fix the `key_field` configuration that is used for a UUID

The latter change is backwards compatible, but gaining the advantage of
the fast UUID field will require a reindex.

## Tests

Added a repro test which now shows `Mixed` execution used for both a
`key_field` UUID, and a normal UUID field.

Regarding backwards compatibility: I manually created the table/index
from the regression test (which has both a UUID `key_field` and a fast
non-`key` UUID field) before this commit, and then upgraded to this
commit. I confirmed that the `key_field` schema was [unchanged
before/after the
upgrade](https://gist.github.com/stuhood/cdf0976909f888645c0269c748392858),
confirming that its schema was persisted. To actually get the benefit of
a `fast` `key_field` for a UUID column, users will need to reindex, but
the other bug fix in this change will take effect without a reindex.