perf: reference count open relations (#2716)
While investigating what happens when an index has many thousands of
segments the profiler indicated that opening the underlying index
Relation over and over throughout the searching codepath accounted for
nearly 20% of the execution time.

This PR introduces a new type called `PgSearchRelation` which looks a
lot like `pgrx::PgRelation` but is clone-able and internally uses
references counting to elide re-opening the relation and also to ensure
the underlying relation is only closed when there are no more references
(and so long as we're still in a transaction).

Trying to cut out search overhead. In a fairly degenerate case of an
index with 9885 segments, a query dropped from 1.3s to 618ms:

```
[v16.2][399740] reddit=# set max_parallel_workers_per_gather = 0; select count(*) from reddit where body @@@ 'beer';
SET
Time: 0.406 ms
 count
-------
 85823
(1 row)
Time: 1355.696 ms (00:01.356)
```
```
[v16.2][400499] reddit=# set max_parallel_workers_per_gather = 0; select count(*) from reddit where body @@@ 'beer';
SET
Time: 0.405 ms
 count
-------
 85823
(1 row)
Time: 618.923 ms
```

Introducing a new type to be our primary interface into opening/managing
a relation and updating all the code paths to use it.

Existing tests pass.

---------

Signed-off-by: Eric Ridge <eebbrr@gmail.com>
Co-authored-by: Ming <ming.ying.nyc@gmail.com>
Co-authored-by: Stu Hood <stuhood@paradedb.com>