chore: cleanup around `SearchIndexSchema::open()` (#2727)
## What

In nearly every case we have enough information to construct a
`SearchIndexSchema` without also needing to instantiate a
`MVCCDirectory` and a tantivy `Index`.

Invent a new function, `SearchIndexSchema::from_index()` to require the
caller to pass the necessary information and update all the old
compatible `SearchIndexSchema::open()` call points.

## Why

General code cleanup and also removes overhead of opening an Index and
MVCCDirectory when we already have them.

## How

## Tests