perf: optimize for `paradedb.all()` in a "Filter:" context (#2729)
## What

When we see that `paradedb.all()` is the query being evaluated when our
`@@@` operator is running in a Postgres "Filter:" context, we
essentially store a flag indicating "all" rather than actually querying
the index and materialzing literally every key_field value into a
HashSet.

## Why

When `@@@` is run via a Postgres "Filter:" node we're asked to
materialize every matching primary key into an in-memory HashSet, and
this is true for every individual predicate using `@@@`.

When one (or heaven forbid, more!) of those predicates is `@@@
paradedb.all()` we'd literally materialize every primary key into a
HashSet for each predicate. And if parallel workers are involved, that's
potentially multiple copies of every primary key per worker process.
This could *quickly* lead to OOM situations.

## How

The power of Rust enums.

## Tests

Indeed.