feat: add `target_segment_count` as an index `WITH` option (#2737)
## What

Adds a new index `WITH` option to let users set the
`target_segment_count`.

It probably makes sense to have this value as a property of the index
itself so that the same value can be reused during REINDEX. If
unspecified it defaults to the number of CPUs on the host system.

After discussions with @rebasedming, we have decided to remove the
`paradedb.target_segment_count` GUC.

## Tests

Existing tests that used the old GUC have been updated to use this new
index option. Docs updated too.