add `MvccSatisfies::LargestSegment` variant
This new variant allows us, during planning, to do the minimal amount of
work necessary to open a tantivy `Index` so we can do cardinality and
selectivity estimation.

There's also a new `MvccDirectory::total_segment_count()` function that
returns a lazily-assigned count of the total number of visible segments
in the index, which may not be the same as the actual number of segments
that were selected for use -- especially in the case of the new
`MvccSatisfies::LargestSegment` variant.

This count is used during our custom scan planning callback.