chore: Quiet down the proptests (#2750)
## What

Remove extraneous debug output from proptests.

## Why

Proptests run many times in a loop: to reduce their cost and the amount
that they output, it is recommended to put any information that would
help debug a failure into a test macro like `prop_assert_eq!`: that way,
it will only actually be computed and rendered for failing cases _after_
the framework has minimized to the smallest repro that it can find.

When debugging the cases that are being run, you can use
`PROPTEST_VERBOSE=3` to have the framework print debug information about
which cases it is trying.