fix: Do not convert vars to external on a base relation (#2754)
# Ticket(s) Closed

- Closes #2753

## What

When extracting quals from a join `RestrictInfo`, we will convert vars
into external vars. But one of the cases of that conversion was not
covered by a check that we were extracting quals for a join.

Additionally, add an enabled-by-default
`paradedb.enable_custom_scan_without_operator` GUC to control whether
the custom scan can be used when we've successfully found expressions to
push down, but have _not_ found use of our operator.

## Why

See #2753.

## Tests

Added the repro from #2753, and added a test of the new GUC.