fix: Revert "optimize `SearchIndexReader::collect_segments()`" (#2757)
## What

This reverts commit 23a697611af94765fd1f3fce1d179345753fdc1c.

## Why

The benchmark suite (and a local bisect) indicated that that commit
(from #2742) roughly doubled the time for TopN.

That was because the `TopN` exec method lazily produces segments by
checking them out one at a time. The effect of eagerly contructing a set
of segments to search with was that all of the segments were very likely
to be checked out by one worker.

## How

Revert 23a697611af94765fd1f3fce1d179345753fdc1c

See #2758 for followup around re-fixing this!