chore: more improvements to index/schema configuration and management (#2771)
## What

#2660 brought a much needed round of cleanups to how we manage index
schemas. Unfortunately, it introduced quite some overhead in
reading/decoding/validating the schema. This process was happening quite
a bit throughout the execution paths of `aminsert` and other hot-spots.

#2176 brought the ability to essentially keep one heavy-weight
`PgSearchRelation` instantiated and cheaply clone it when necessary.
This PR cleans up things further such that the `SearchIndexSchema` is
now a lazily-evaluated property of `PgSearchRelation`. This means
`SearchIndexSchema` is only evaluated when needed, and then only once
(at least per statement).

Furthermore, its internal properties are lazily-evaluated, ensuring any
given code path doesn't do more work than it needs.

This also renames `SearchIndexOptions` to `BM25IndexOptions`, mainly
because I kept getting confused about what `SearchIndexOptions`
represented (it was too similarly named to `SearchIndexSchema` for my
tastes). And `BM25IndexOptions` is now a property of `PgSearchRelation`
too.

This seems to have drastically improved the write throughput of the
INSERT/UPDATE jobs in our `single-server.toml` stressgress test.
v0.15.26 was 176/s INSERTs and 154/s UPDATEs. This PR clocks in at 275/s
and 260/s, respectively.

# Other Notable Changes

- Index configuration validation now happens during CREATE INDEX/REINDEX
in `ambuildempty()` rather than on every instantiation of
`SearchIndexSchema`.

- The "raw" tokenizer deprecation warnings are now gone, unless somehow
the "key_field" is configured with it -- which is no longer possible

## Why

Trying to rollback performance regressions that were introduced in
0.16.0

## How

## Tests

All existing tests pass, and a few were updated due to the "raw"
tokenizer deprecation warning going away and a change in wording for a
specific validation error.