feat: Update release process to follow a singular branch (#2721)
- Closes #N/A

This PR changes our repository branching model to work off of a single
branch, `dev`, from which we manually trigger the creation of releases.
The entire logic is documented in `RELEASE.md`, so I'm not going to
rehash it here.

We need the ability to make hotfixes to previous releases.

See RELEASE.md

The CI here tests that the general development flow still works. As for
the deployment flow, I tested various sub-steps locally, but a full
end-to-end version needs to be done. That's why I added the `dryrun`
input, which I'll use to do a test once we merge this to `dev`.

---------

Signed-off-by: Philippe Noël <21990816+philippemnoel@users.noreply.github.com>
Co-authored-by: Stu Hood <stuhood@paradedb.com>