ci: disable using fsm_info() in stressgres suite (#2803)
For now disable using fsm_info() in stressgres `bulkd-updates.toml`
suite.

This is because the benchmark workflows use the latest suite files from
`main` to run against prior branches during a backfill and this is a
relatively new function so old branches don't have it.

I am not sure how to handle this going forward. Perhaps stressgres can
be taught how to ignore certain errors -- I am not sure.

---------

Signed-off-by: Eric Ridge <eebbrr@gmail.com>