ci: sleep for random seconds before publishing to gh-pages (#2813)
It's possible for concurrent jobs that were spawned at the same time via
a benchmark backfill to try and publish the benchmark results to
gh-pages at the same time.

While the `benchmark-action` has some accounting for this with a
hard-coded retry count (of 10), it's not necessarily enough, so we
introduce a random sleep between 1 and 66 seconds, in an effort to fend
this problem off.