ci: Manually deploy Mintlify (#2787)
# Ticket(s) Closed

- Closes #N/A

## What
Now that we moved to a single branch workflow, we need Mintlify docs to
be deployed at the same time as the manually triggered releases, rather
than automatically on pushes to `main`. This does that.

We need to:
- [x] Populate the GitHub Secret for the API key
- [x] Disable the automated deployment

I need @rebasedming's help for that, since I don't have access. EDIT:
Done, thank you Ming!

## Why
^

## How
^

## Tests
Need to test in CI once the above TODOs are handled.