chore: Rename datasets and add string paging queries (#2834)
## What

Add a high-cardinality paging/top-n query to the benchmarks, and rename
datasets to match their content. Additionally, improve the generation
script for the `docs` dataset to avoid joins and allow for deterministic
relative-position queries.

## Why

We don't currently have a high-cardinality string paging/top-n query in
the benchmark. We have top-n on a string column, but only for
low-cardinality values (`top_n-string.sql`). The top-n case represented
an important gap that a user encountered, which #2828 addresses.

The names of the `benchmark` datasets don't currently describe their
shape / schema, and for the `join` dataset in particular, that would
discourage using it for other types of queries. We rename it to `docs`
here, and then use the `pages` table as the dataset for top-n.

## Tests

Tested locally that the new query demonstrates a speedup for #2828.