fix: orphaned delete entries get GCed too early (#2845)
# Ticket(s) Closed

- Closes #

## What

When running a new stressgres suite (coming in a future PR), I hit a
mysterious bug where it looked like vacuum could cause corruption of
some pages.

Turns out it's caused by scenarios where:

1. A `DeleteEntry` already exists for a `SegmentMetaEntry`, and a new
one is created
2. A new, "fake" `SegmentMetaEntry` gets created for the purpose of
storing the old `DeleteEntry`, so its blocks can get garbage collected
3. Because this "fake" entry is invisible to all readers besides the
garbage collector, it doesn't get pinned and can get garbage collected
too early (i.e. while a reader is still pinning the old `DeleteEntry`)

The solution is to copy all of the contents of the old
`SegmentMetaEntry` to the fake one, so that the "pintest blockno" of the
fake entry is that same as that of the entry with the new `DeleteEntry`.
That way, the `DeleteEntry` doesn't get garbage collected until the pin
is released.

## Why

## How

## Tests