ci: Mark release as latest if semVer is higher (#2867)
# Ticket(s) Closed

- Closes #N/A

## What
We noticed that releasing v0.16.5 as a patch didn't mark as `latest` in
GitHub. This is because we only marked releases from `main` as latest,
but in the edge case where we do a patch release of an ongoing latest
release, we still need to mark as latest.

This PR does this by using the `semver` NPM package to compare versions
in a semver-compliant way. I've tested that it works, and that it will
work with `-rc.` format as well. Only official patch or minor versions
get marked latest, never `-rc.`.

0.16.5 > 0.16.4 no matter what branch it is run on
0.17.0 > 0.16.4 no matter what branch it is run on
0.16.5 !> 0.16.4 no matter what branch it is run on (not marked latest)

## Why
Working releases

## How
^

## Tests
Manually tested since we don't run the release before, well, we release!

---------

Signed-off-by: Philippe Noël <21990816+philippemnoel@users.noreply.github.com>