fix: issue #2844 - Use `paradedb.with_index()` when necessary
When the right-hand-side of the `@@@` operator is a complex expression
that can't be evaluated to a `T_Const` during planning, we need to
rewrite it to be a call to our `paradedb.with_index(<indexrelid>,
<rhs>)` so that it can be properly used in Filter contexts during
execution.

This seems to come up, at least, during RLS rule evaluation if those
rules use our `@@@` operator.