fix: issue #2844 - Use `paradedb.with_index()` when necessary (#2879)
# Ticket(s) Closed

- Closes #2844

## What

When the right-hand-side of the `@@@` operator is a complex expression
that can't be evaluated to a `T_Const` during planning, we need to
rewrite it to be a call to our `paradedb.with_index(<indexrelid>,
<rhs>)` so that it can be properly used in Filter contexts during
execution.

This seems to come up, at least, during RLS rule evaluation if those
rules use our `@@@` operator.

## Why

So long as the right-hand-side of `@@@` is a `SearchQueryInput`, we need
to be able to execute it when Postgres wants us to, whether that's
during a scan or, as in this case, a filter being run as part of RLS
evaluation.

## How

Wrapping Postgres expression nodes!

## Tests

I used the re-create the OP from #2844 put together as a regression
test. I seem to be lacking the creative energy today to reduce the test
down to its simplest form that doesn't require a complicated RLS
configuration.