perf: Avoid a re-query when a top-n query is exhausted (#2888)
## What

Split the iterators consumed by the `TopN` exec method and by the
`Normal`/`FF` exec methods into two separate implementations.

After splitting them, use the known exact size of the `TopN` iterator to
exit early (as `exhausted: bool`) if we had fewer matches than were
requested. This avoids re-querying an iterator (and re-scanning its
columns) that we know has no more results.

## Why

The two types of consumers of these iterators have very different
consumption patterns and constraints:
* TopN knows exactly how many results there will be, and already buffers
them all in memory. It would like to be able to know the precise count
of results.
* `Normal` and `FF` exec methods do not know the total number of results
to expect, and in some cases would like to be able to consume in a
segment-aware fashion in order to be able to late-fetch fast field
columns (see #2623).

## Tests

Added a test to cement the change in query count.

Benchmarks show a 1.5x speedup for `paging-string-max`, and no change
for other queries.