feat: text search operators (#2877)
## What

This is the implementation of 4 new operators designed around common
"text search" patterns: `===` (exact token), `&&&` (tokenized term
conjunction) `|||` (tokenized term disjunction), and `###` (phrase).

First, `@@@` still exists and still does what it's always done, which is
run the right-hand-side through tantivy's "query parser".

These new operators have specific meanings. The spelling of these
operators is both to maintain some kind of consistency in that they're
all 3 of the same symbol and also to try and confer their meaning.

In all cases, the left-hand-side of the operator is required to be a
reference to a Postgres `TEXT`/`VARCHAR` field and the right-hand-side
is text. The operator definition defines how the right-hand-side input
is applied to the left-hand-side field. These are "text search"
operators and would be nonsensical applied to non-text fields.

- `field === 'TEXT'` - this is the exact term match operator. The
right-hand-side value is used, unmodified, to perform the equivalent of
the `paradedb.term()` function.

- `field === ARRAY['...', '...', '...']` - similar to the above, but
rewrites to the `paradedb.term_set()` function.

- `field &&& 'TEXT'` - this is the "match conjunction" operator. It
tokenizes the right-hand-side using the index tokenizer defined by the
left-hand-side field and searches for documents that contain **all** of
the tokenized terms. Think, "find all documents that contain all terms
tokenized from this text input". This is the equivalent of the
`paradedb.match` function with the `conjunction_mode => true`.

- `field ||| 'TEXT'` - this is the "match disjunction" operator. It
tokenizes the right-hand-side using the index tokenizer defined by the
left-hand-side field and searches for documents that contain **any** of
the tokenized terms. Think, "find all documents that contain one or more
of the terms tokenized from this text input". This is the equivalent of
the `paradedb.match` function with the `conjunction_mode => false`.

- `field ### 'TEXT'` - this is the "phrase search" operator. It
tokenizes the right-hand-side using the index tokenizer defined by the
left-hand-side field and searches for documents that contain all the
tokenized terms in the order written. Think, "find all documents that
contain this tokenized phrase". This is akin to the `paradedb.phrase()`
function, except the `###` is transparently handling the tokenization.

If you wanted to find all documents that talk about bbq and chicken and
maybe hotdogs and happen on July 4th, you might write a query like this:

```sql
SELECT * 
FROM events 
WHERE (
       description &&& 'bbq chicken' OR description ||| 'hot dog hotdogs'
) AND description ### 'July 4th';
```

## Why

These operators are in furtherance of our
[roadmap](https://github.com/orgs/paradedb/discussions/2041) --
specifically the "Database UX improvements" line item.

Our overall goal with that roadmap item is to start moving users and ORM
tools away from needing to use our "builder functions" in the common
text-search cases. What we're striving for is for users, and ORM tools,
to write standard SQL WHERE clauses. In order to do that we need to make
common things more accessible using custom operators.

## How

Largely just adding new `#[pg_operator]` functions and ensuring they all
have a SUPPORT function that can rewrite the expression to what we need.

@rebasedming said he was going to tackle documentation, which can come
in a separate PR.

## Tests

There's a new regression test named `operators.sql` that validates
EXPLAIN output and search results against our standard "mock_items"
table.