perf: Lazily load bitpacked columns (#2894)
## What

Incorporates https://github.com/paradedb/tantivy/pull/56.

## Why

As mentioned there:
> We would like to be able to lazily load `BitpackedCodec` columns
(similar to what
https://github.com/paradedb/tantivy/commit/020bdffd61365a140218643c49ba01c5043b2966
did for `BlockwiseLinearCodec`), because in the context of `pg_search`,
immediately constructing `OwnedBytes` means copying the entire content
of the column into memory.

## Tests

There are a few 2x speedups in the benchmark suite, as well as a 1.8x
speedup on a representative customer query.

Unfortunately there are also some 13-19% slowdowns on aggregates with
`solve_mvcc=false`: it looks like that is because aggregates use
`get_vals`, for which the default implementation is to just call
`get_val` in a loop. After discussion, we think that getting back that
performance might require wider API changes to make batching more
inherent.